requirements.txt
# Core
aiogram>=3.4.1
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.27
aiosqlite>=0.19.0
alembic>=1.13.1

# HTTP Client
httpx>=0.26.0

# Utils
python-dotenv>=1.0.1
pytz>=2024.1
uuid>=1.30
